removed enum.py from itom-packages since it causes compatibility issues with the 'real' enum.py of the default python installation.
Background: The module enum.py is part of Python since Python 3.4. In order to also use enum functionality in Python <= 3.3 a copy
of the future enum.py module was added to itom-packages. However, newer Python versions have an extended version of enum.py, such that
Pip or other packages will fail if the 'old' copy of enum.py in the itom-packages directory is used. Therefore, enum.py is removed from
itom. If one uses Python < 3.4, the package enum34 has to be installed in order to get the same functionality, too.