workaround in font selection of WidgetPropEditorStyles:
If the current language of itom does not correspond to the OS language (see with Windows),
the style list in the dialog is translated to the OS language, however the styleName of the
font is in the itom language (e.g. english). Then, the styles are not correctly selected.
Avoid this by testing all available styles of the font family and compare the resulting font with
the given font.