cleanup of constructor of itom.dataObject. Added further dataObject related unittests in python_unittests to show the desired behaviour of all kind of constructors / copy-constructors of itom.dataObject. The copy constructor now considers an optional dtype or continuous argument. All arguments can now be keyword based.
Currently, itom.dataObject has a wrong behaviour for boolean comparisons (dataObject == True always returns True).
This will be corrected in the future. Up to now, a deprecation warning is raised to show that this method will
be fixed in the future. It is important to know, that code like 'if myDataObject:' will not be automatically True in
the future.