Switched CMAKE to version 3.12 to use CMAKE_PROJECT_VERSION feature. Changed way to define project Version. Now global.h is configured via CMAKE as global.h. in and the stored in the Build directory as global.h.
This should simplify version definition and reduced the level of redundancy.

At this point it became evident that many of the local include definitions are
redudant with the CMAKE inlcude dir directive. Hence also the the include definition
for all files have been changed accordingly.

At a later stage it is highly recommended to set the Project Version
via the GIT Tag to further reduce redundancy an potential errror source.